+++ Source/Models/BrickModels.swift
@@ -46,7 +46,7 @@ public class Brick: CustomStringConvertible {
     public var backgroundView: UIView?
     
     /// Datasource used to handle which view to show for state. Defaults to nil
-    public var brickCellAppearanceDataSource: BrickCellAppearanceDataSource?
+    public weak var brickCellAppearanceDataSource: BrickCellAppearanceDataSource?
     
     /// Delegate used to handle tap gestures for the brick. Defaults to nil
     public weak var brickCellTapDelegate: BrickCellTapDelegate?
